flask/requirements.txt
annotated-types==0.6.0
blinker==1.6.3
click==8.1.7
dnspython==2.4.2
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mongoengine==0.27.0
pydantic==2.4.2
pydantic_core==2.10.1
pymongo==4.5.0
python-dotenv==1.0.0
typing_extensions==4.8.0
Werkzeug==3.0.1
